5/31까지 취합본

+++ src/main/java/com/ducks/demys/boot/vo/ProjectsForPrint.java
@@ -28,7 +28,7 @@ public class ProjectsForPrint {
 	
 	private String MEMBER_NAME;
 	private String CT_NAME;
-	private Double SUM_PERCENT;
+	private int SUM_PERCENT;
 	private int ISSUE_COUNT;
 	
 	private String SORT;
